# [generator] Enable parallel type generation.