#446 removed Cecil usage from the step that generates the type `.cs` files, which was the remaining thread-safety issue.  This PR enables generating those files in parallel.

This results in a nice performance win, particularly for projects binding a lot of types:

Scenario | Original | This PR
------------ | ------------- | -
Mono.Android.dll | 16423ms | 11506ms
Jar2Xml | 5517ms | 5309ms
Class-Parse | 15511ms | 15615ms