@jonpryor 2019-07-18
Odd that `class-parse` got *slower*...