@jonpryor 2019-07-18
What have you done in terms of "stress testing" to ensure that PR #446 didn't "miss" anything?  Have you tried running `generator.exe` multiple times into separate output directories, and `diff` the generated outputs to ensure consistency?