@jpobst 2019-07-18
I shouldn't have included `class-parse` results for this PR, as that test case does not write types.  (it uses `--only-xml-adjuster`)  It's just margin of error.  ;)

My testing is roughly what you suggest.  I committed the "known good" output to a local git repository.  I then run `generator.exe` into a new directory, `robocopy /purge` it to the git directory (so it would remove any missing files), and then git diff it to ensure nothing has changed.